# [xamlc] return early in `NodeILExtensions` for `ValidateOnly`