--- a/src/Controls/src/Build.Tasks/NodeILExtensions.cs
+++ b/src/Controls/src/Build.Tasks/NodeILExtensions.cs
@@ -422,6 +422,10 @@ namespace Microsoft.Maui.Controls.Build.Tasks
 
 		public static IEnumerable<Instruction> PushXmlLineInfo(this INode node, ILContext context)
 		{
+			if (context.ValidateOnly)
+			{
+				yield break;
+			}
 			var module = context.Body.Method.Module;
 
 			var xmlLineInfo = node as IXmlLineInfo;
@@ -447,6 +451,10 @@ namespace Microsoft.Maui.Controls.Build.Tasks
 
 		public static IEnumerable<Instruction> PushParentObjectsArray(this INode node, ILContext context)
 		{
+			if (context.ValidateOnly)
+			{
+				yield break;
+			}
 			var module = context.Body.Method.Module;
 
 			var nodes = new List<IElementNode>();
@@ -585,6 +593,10 @@ namespace Microsoft.Maui.Controls.Build.Tasks
 
 		public static IEnumerable<Instruction> PushServiceProvider(this INode node, ILContext context, FieldReference bpRef = null, PropertyReference propertyRef = null, TypeReference declaringTypeReference = null)
 		{
+			if (context.ValidateOnly)
+			{
+				yield break;
+			}
 			var module = context.Body.Method.Module;
 
 #if NOSERVICEPROVIDER